Move a few helper methods into files with clearer names
