+++ app/helpers/shiny_main_site_helper.rb
@@ -6,20 +6,14 @@
 #
 # ShinyCMS is free software; you can redistribute it and/or modify it under the terms of the GPL (version 2 or later)
 
-# Methods that might be useful in templates and/or controllers on the main site
+# Includes all of the helpers that might be useful in templates and/or controllers on the main site
 module ShinyMainSiteHelper
+  include ShinyConsentHelper
   include ShinyDiscussionHelper
   include ShinyPluginHelper
+  include ShinySettingsHelper
   include ShinySiteNameHelper
   include ShinyUserHelper
 
   include ActsAsTaggableOn::TagsHelper
-
-  def consent_version( slug )
-    ConsentVersion.find_by( slug: slug )
-  end
-
-  def setting( name )
-    Setting.get( name, current_user )
-  end
 end
