+++ app/helpers/shiny_plugin_helper.rb
@@ -19,4 +19,8 @@ module ShinyPluginHelper
   def plugins_with_admin_toolbar_templates
     ShinyPlugin.with_template( 'admin/toolbar/_section.html.erb' )
   end
+
+  def valid_plugin_path?( plugin_path = nil )
+    plugin_path.present? && plugin_path.starts_with?( 'plugins/' ) && plugin_path.ends_with?( 'app/views' )
+  end
 end
