+++ app/helpers/shiny_settings_helper.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+# ShinyCMS ~ https://shinycms.org
+#
+# Copyright 2009-2020 Denny de la Haye ~ https://denny.me
+#
+# ShinyCMS is free software; you can redistribute it and/or modify it under the terms of the GPL (version 2 or later)
+
+# Helper method for retrieving site settings
+module ShinySettingsHelper
+  def setting( name )
+    Setting.get( name, current_user )
+  end
+end
